Move helpers into ./utils directory
