+++ sdk/containerregistry/container-registry/src/utils/base64.browser.ts
