+++ sdk/containerregistry/container-registry/src/utils/wwwAuthenticateParser.ts
